# contracts/openapi/powerflow.openapi.json
{
  "components": {
    "schemas": {
...
      "BatteryConfig": {
        "additionalProperties": false,
        "properties": {
          "aux_load_kw": {
            "default": 0.0,
            "description": "Auxiliary load, drawn from the LV bus (not the cells).",
            "minimum": 0.0,
            "title": "Aux Load Kw",
            "type": "number"
          },
          "capacity_kwh": {
            "exclusiveMinimum": 0.0,
            "title": "Capacity Kwh",
            "type": "number"
          },
          "efficiency": {
            "$ref": "#/components/schemas/EfficiencyConfig"
          },
          "soc_initial_pct": {
            "default": 50.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Soc Initial Pct",
            "type": "number"
          },
          "soc_max_pct": {
            "default": 95.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Soc Max Pct",
            "type": "number"
          },
          "soc_min_pct": {
            "default": 5.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Soc Min Pct",
            "type": "number"
          }
        },
        "required": [
          "capacity_kwh"
        ],
        "title": "BatteryConfig",
        "type": "object"
      },
      "BessAssetResponse": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/BessConfig"
          },
          "measurement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BessMeasurement"
              },
              {
                "type": "null"
              }
            ],
            "description": "null before the first step."
          },
          "setpoint": {
            "$ref": "#/components/schemas/BessSetpointState"
          }
        },
        "required": [
          "config",
          "setpoint",
          "measurement"
        ],
        "title": "BessAssetResponse",
        "type": "object"
      },
      "BessConfig": {
        "additionalProperties": false,
        "properties": {
          "battery": {
            "$ref": "#/components/schemas/BatteryConfig"
          },
          "collector": {
            "default": "mv1",
            "title": "Collector",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Id",
            "type": "string"
          },
          "inverter": {
            "$ref": "#/components/schemas/BessInverterConfig"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "transformer": {
            "$ref": "#/components/schemas/TransformerConfig"
          }
        },
        "required": [
          "id",
          "inverter",
          "battery",
          "transformer"
        ],
        "title": "BessConfig",
        "type": "object"
      },
...
      "BessMeasurement": {
        "description": "Generator convention: P > 0 discharging.",
        "properties": {
          "alarm_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Alarm Flag Names",
            "type": "array"
          },
          "alarm_flags": {
            "title": "Alarm Flags",
            "type": "integer"
          },
          "aux_p_kw": {
            "title": "Aux P Kw",
            "type": "number"
          },
          "energy_available_charge_kwh": {
            "title": "Energy Available Charge Kwh",
            "type": "number"
          },
          "energy_available_discharge_kwh": {
            "title": "Energy Available Discharge Kwh",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "limit_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Limit Flag Names",
            "type": "array"
          },
          "limit_flags": {
            "title": "Limit Flags",
            "type": "integer"
          },
          "operating_state": {
            "title": "Operating State",
            "type": "integer"
          },
          "operating_state_name": {
            "title": "Operating State Name",
            "type": "string"
          },
          "p_available_charge_kw": {
            "title": "P Available Charge Kw",
            "type": "number"
          },
          "p_available_discharge_kw": {
            "title": "P Available Discharge Kw",
            "type": "number"
          },
          "p_cmd_kw": {
            "title": "P Cmd Kw",
            "type": "number"
          },
          "p_kw": {
            "title": "P Kw",
            "type": "number"
          },
          "q_cmd_kvar": {
            "title": "Q Cmd Kvar",
            "type": "number"
          },
          "q_kvar": {
            "title": "Q Kvar",
            "type": "number"
          },
          "s_kva": {
            "title": "S Kva",
            "type": "number"
          },
          "soc_pct": {
            "title": "Soc Pct",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "status_name": {
            "title": "Status Name",
            "type": "string"
          },
          "v_lv_kv": {
            "title": "V Lv Kv",
            "type": "number"
          },
          "v_lv_pu": {
            "title": "V Lv Pu",
            "type": "number"
          }
        },
        "required": [
          "id",
          "p_cmd_kw",
          "q_cmd_kvar",
          "p_kw",
          "q_kvar",
          "s_kva",
          "soc_pct",
          "energy_available_discharge_kwh",
          "energy_available_charge_kwh",
          "p_available_discharge_kw",
          "p_available_charge_kw",
          "status",
          "status_name",
          "limit_flags",
          "limit_flag_names",
          "aux_p_kw",
          "v_lv_pu",
          "v_lv_kv",
          "operating_state",
          "operating_state_name",
          "alarm_flags",
          "alarm_flag_names"
        ],
        "title": "BessMeasurement",
        "type": "object"
      },
      "BessMode": {
        "enum": [
          "idle",
          "pq",
          "offline"
        ],
        "title": "BessMode",
        "type": "string"
      },
      "BessSetpointRequest": {
        "additionalProperties": false,
        "description": "A partial BESS setpoint: omitted fields keep their current value.",
        "properties": {
          "mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BessMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "p_kw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "+ discharge, − charge (kW).",
            "title": "P Kw"
          },
          "q_kvar": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "+ injecting vars (kvar).",
            "title": "Q Kvar"
          }
        },
        "title": "BessSetpointRequest",
        "type": "object"
      },
      "BessSetpointState": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/BessMode"
          },
          "p_kw": {
            "title": "P Kw",
            "type": "number"
          },
          "q_kvar": {
            "title": "Q Kvar",
            "type": "number"
          }
        },
        "required": [
          "p_kw",
          "q_kvar",
          "mode"
        ],
        "title": "BessSetpointState",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoadAssetResponse": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/LoadConfig"
          },
          "measurement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LoadMeasurement"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "config",
          "measurement"
        ],
        "title": "LoadAssetResponse",
        "type": "object"
      },
      "LoadConfig": {
        "additionalProperties": false,
        "properties": {
          "bus": {
            "default": "poi",
            "description": "\"poi\" or a collector id.",
            "title": "Bus",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "noise": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NoiseConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileRef"
          },
          "transformer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformerConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional transformer; its LV side feeds the load."
          }
        },
        "required": [
          "id",
          "profile"
        ],
        "title": "LoadConfig",
        "type": "object"
      },
      "LoadMeasurement": {
        "description": "Load convention: P > 0 consuming.",
        "properties": {
          "alarm_flag_names": {
            "items": {
              "type": "string"
            },
            "title": "Alarm Flag Names",
            "type": "array"
          },
          "alarm_flags": {
            "title": "Alarm Flags",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "p_kw": {
            "title": "P Kw",
            "type": "number"
          },
          "pf": {
            "title": "Pf",
            "type": "number"
          },
          "q_kvar": {
            "title": "Q Kvar",
            "type": "number"
          },
          "s_kva": {
            "title": "S Kva",
            "type": "number"
          },
          "supply_state": {
            "title": "Supply State",
            "type": "integer"
          },
          "supply_state_name": {
            "title": "Supply State Name",
            "type": "string"
          },
          "v_pu": {
            "title": "V Pu",
            "type": "number"
          }
        },
        "required": [
          "id",
          "p_kw",
          "q_kvar",
          "s_kva",
          "pf",
          "v_pu",
          "supply_state",
          "supply_state_name",
          "alarm_flags",
          "alarm_flag_names"
        ],
        "title": "LoadMeasurement",
        "type": "object"
      },
...
      "NoiseConfig": {
        "additionalProperties": false,
        "properties": {
          "p_std_pct": {
            "default": 0.0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "P Std Pct",
            "type": "number"
          },
          "q_std_pct": {
            "default": 0.0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Q Std Pct",
            "type": "number"
          }
        },
        "title": "NoiseConfig",
        "type": "object"
      },
...
      "PointValueResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "name",
          "value",
          "unit"
        ],
        "title": "PointValueResponse",
        "type": "object"
      },
...
      "PvAssetResponse": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/PvConfig"
          },
          "measurement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PvMeasurement"
              },
              {
                "type": "null"
              }
            ]
          },
          "setpoint": {
            "$ref": "#/components/schemas/PvSetpointState"
          }
        },
        "required": [
          "config",
          "setpoint",
          "measurement"
        ],
        "title": "PvAssetResponse",
        "type": "object"
      },
...
      "PvConfig": {
        "additionalProperties": false,
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/PvAvailabilityConfig"
          },
          "collector": {
            "default": "mv1",
            "title": "Collector",
            "type": "string"
          },
          "dc_kwp": {
            "exclusiveMinimum": 0.0,
            "title": "Dc Kwp",
            "type": "number"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Id",
            "type": "string"
          },
          "inverter": {
            "$ref": "#/components/schemas/PvInverterConfig"
          },
          "loss_factor": {
            "default": 0.0,
            "description": "Fraction lost before AC.",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Loss Factor",
            "type": "number"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "transformer": {
            "$ref": "#/components/schemas/TransformerConfig"
          }
        },
        "required": [
          "id",
          "dc_kwp",
          "inverter",
          "availability",
          "transformer"
        ],
        "title": "PvConfig",
        "type": "object"
      },
...
      "PvSetpointRequest": {
        "additionalProperties": false,
        "description": "A partial PV setpoint: at most one of p_limit_kw / p_limit_pct and one of q_kvar / pf.\nOmitted fields keep their current value.",
        "properties": {
          "p_limit_kw": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Curtailment limit (kW).",
            "title": "P Limit Kw"
          },
          "p_limit_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Curtailment limit (% of p_max_kw).",
            "title": "P Limit Pct"
          },
          "pf": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": -1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Power factor; + injecting vars, − absorbing; |pf| ≥ 0.8. Selects PF mode.",
            "title": "Pf"
          },
          "q_kvar": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Q setpoint; selects Q mode.",
            "title": "Q Kvar"
          }
        },
        "title": "PvSetpointRequest",
        "type": "object"
      },
      "PvSetpointState": {
        "properties": {
          "p_limit_kw": {
            "title": "P Limit Kw",
            "type": "number"
          },
          "p_limit_pct": {
            "title": "P Limit Pct",
            "type": "number"
          },
          "pf": {
            "title": "Pf",
            "type": "number"
          },
          "q_kvar": {
            "title": "Q Kvar",
            "type": "number"
          },
          "q_mode": {
            "description": "\"q\" or \"pf\".",
            "title": "Q Mode",
            "type": "string"
          }
        },
        "required": [
          "p_limit_kw",
          "p_limit_pct",
          "q_mode",
          "q_kvar",
          "pf"
        ],
        "title": "PvSetpointState",
        "type": "object"
      },
...
      "SetpointResult": {
        "properties": {
          "accepted": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BessSetpointState"
              },
              {
                "$ref": "#/components/schemas/PvSetpointState"
              }
            ],
            "title": "Accepted"
          },
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "asset_type": {
            "title": "Asset Type",
            "type": "string"
          },
          "clamped": {
            "title": "Clamped",
            "type": "boolean"
          },
          "flags": {
            "description": "Static limits that clamped the request.",
            "items": {
              "type": "string"
            },
            "title": "Flags",
            "type": "array"
          },
          "requested": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Requested",
            "type": "object"
          }
        },
        "required": [
          "asset_type",
          "asset_id",
          "requested",
          "accepted",
          "clamped",
          "flags"
        ],
        "title": "SetpointResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Grid-connected microgrid power flow simulator (pandapower, balanced positive sequence,\nquasi-static). The EMS writes setpoints; every step the simulator applies the asset limits,\nsolves the network and publishes a measurement snapshot.\n\n**Sign conventions**\n- BESS / PV P and Q: generator convention. P > 0 injects (BESS discharging, PV producing),\n  P < 0 is a charging BESS. Q > 0 injects vars.\n- Load P and Q: P > 0 consumes, Q > 0 consumes vars (lagging).\n- POI: P > 0 exports to the utility, P < 0 imports. The same for Q.\n- Transformers: P/Q > 0 flows toward the grid (LV → HV).\n- pf = |P|/S, signed with Q in the element's own convention.\n\n**Units:** kW, kvar, kVA, kWh, kV (line-to-line), pu where noted, s.\n\n**Points:** `<asset_type>.<asset_id>.<point>`, listed at GET /api/points.\n",
    "title": "powerflow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/assets/bess/{asset_id}": {
      "get": {
        "operationId": "get_bess_api_assets_bess__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BessAssetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown asset or point."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bess",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/bess/{asset_id}/setpoint": {
      "put": {
        "description": "Partial update: omitted fields keep their value. Invalid values → 422; out-of-range values are clamped to the static limits (P ratings, S circle) and reported in `flags`. SOC and ramp limits apply every step and show up in the measurements' `limit_flags`.",
        "operationId": "put_bess_setpoint_api_assets_bess__asset_id__setpoint_put",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BessSetpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetpointResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown asset or point."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set a BESS's P/Q/mode (+P discharge, −P charge)",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/load/{asset_id}": {
      "get": {
        "operationId": "get_load_api_assets_load__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadAssetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown asset or point."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Load",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/pv/{asset_id}": {
      "get": {
        "operationId": "get_pv_api_assets_pv__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PvAssetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown asset or point."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pv",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/pv/{asset_id}/setpoint": {
      "put": {
        "description": "Partial update: omitted fields keep their value. Invalid values → 422; out-of-range values are clamped to the static limits (P ratings, S circle) and reported in `flags`. SOC and ramp limits apply every step and show up in the measurements' `limit_flags`.",
        "operationId": "put_pv_setpoint_api_assets_pv__asset_id__setpoint_put",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PvSetpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetpointResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown asset or point."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set a PV's curtailment (kW or %) and Q or power factor",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/points/{name}": {
      "get": {
        "operationId": "read_point_api_points__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointValueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown asset or point."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one point by name, e.g. bess.bess1.soc_pct",
        "tags": [
          "assets"
        ]
      }
    },
...
  }
}